requirements.txt
annotated-types==0.5.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
docopt==0.6.2
greenlet==2.0.2
idna==3.4
inflect==7.0.0
numpy==1.24.3
pandas==2.0.1
pipreqs==0.4.13
psycopg2==2.9.7
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyMySQL==1.1.0
PyQt6==6.5.0
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.2
xlrd==2.0.1
yarg==0.1.9
